renaming raids and undoing deletion of raid1*
