+++ subjects/quadverifier/audit/README.md
@@ -1,0 +1,73 @@
+#### Quadrangle verifier
+
+##### Try running the program: `"./quadA 3 3 | ./quadVerifier"`
+
+```
+[quadA] [3] [3]
+```
+
+###### Does the program returns the value above?
+
+##### Try running the program: `"./quadB 3 3 | ./quadVerifier"`
+
+```
+[quadB] [3] [3]
+```
+
+###### Does the program returns the value above?
+
+##### Try running the program: `"./quadC 1 1 | ./quadVerifier"`
+
+```
+[quadC] [1] [1] || [quadD] [1] [1] || [quadE] [1] [1]
+```
+
+###### Does the program returns the value above?
+
+##### Try running the program: `"./quadE 1 2 | ./quadVerifier"`
+
+```
+[quadC] [1] [2] || [quadE] [1] [2]
+```
+
+###### Does the program returns the value above?
+
+##### Try running the program: `"./quadE 2 1 | ./quadVerifier"`
+
+```
+[quadD] [2] [1] || [quadE] [2] [1]
+```
+
+###### Does the program returns the value above?
+
+##### Try running the program: `"./quadC 2 1 | ./quadVerifier"`
+
+```
+[quadC] [2] [1]
+```
+
+###### Does the program returns the value above?
+
+##### Try running the program: `"./quadD 1 2 | ./quadVerifier"`
+
+```
+[quadD] [1] [2]
+```
+
+###### Does the program returns the value above?
+
+##### Try running the program: `"./quadE 1 1 | ./quadVerifier"`
+
+```
+[quadC] [1] [1] || [quadD] [1] [1] || [quadE] [1] [1]
+```
+
+###### Does the program returns the value above?
+
+##### Try running the program: `"echo 0 0 | ./quadVerifier"`
+
+```
+Not a Raid function
+```
+
+###### Does the program returns the value above?
