--- subjects/raid1/README.md
@@ -1,569 +1,0 @@
-# raid1
-
-## raid1a
-
-### Instructions
-
-Write a function `Raid1a` that prints a **valid** rectangle of width `x` and of height `y`.
-
-The function must draw the rectangles as in the examples.
-
-`x` and `y` will always be positive numbers. Otherwise, the function should print nothing.
-
-### Expected function
-
-```go
-func Raid1a(x,y int) {
-
-}
-```
-
-### Usage
-
-Here are possible programs to test your function :
-
-Program #1
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1a(5,3)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-o---o
-|   |
-o---o
-student@ubuntu:~/[[ROOT]]/test$
-```
-
-Program #2
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1a(5,1)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-o---o
-student@ubuntu:~/[[ROOT]]/test$
-```
-
-Program #3
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1a(1,1)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-o
-student@ubuntu:~/[[ROOT]]/test$
-```
-
-Program #4
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1a(1,5)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-o
-|
-|
-|
-o
-student@ubuntu:~/[[ROOT]]/test$
-```
-
-----
-
-## raid1b
-
-### Instructions
-
-Write a function `Raid1b` that prints a **valid** rectangle of width `x` and of height `y`.
-
-The function must draw the rectangles as in the examples.
-
-`x` and `y` will always be positive numbers. Otherwise, the function should print nothing.
-
-### Expected function
-
-```go
-func Raid1b(x,y int) {
-
-}
-```
-
-### Usage
-
-Here are possible programs to test your function :
-
-Program #1
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1b(5,3)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-/***\
-*   *
-\***/
-student@ubuntu:~/[[ROOT]]/test$
-```
-
-Program #2
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1b(5,1)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-/***\
-student@ubuntu:~/[[ROOT]]/test$
-```
-
-Program #3
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1b(1,1)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-/
-student@ubuntu:~/[[ROOT]]/test$
-```
-
-Program #4
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1b(1,5)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-/
-*
-*
-*
-\
-student@ubuntu:~/[[ROOT]]/test$
-```
-
-----
-
-## raid1c
-
-### Instructions
-
-Write a function `Raid1c` that prints a **valid** rectangle of width `x` and of height `y`.
-
-The function must draw the rectangles as in the examples.
-
-`x` and `y` will always be positive numbers. Otherwise, the function should print nothing.
-
-### Expected function
-
-```go
-func Raid1c(x,y int) {
-
-}
-```
-
-### Usage
-
-Here are possible programs to test your function :
-
-Program #1
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1c(5,3)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-ABBBA
-B   B
-CBBBC
-student@ubuntu:~/[[ROOT]]/test$
-```
-
-Program #2
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1c(5,1)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-ABBBA
-student@ubuntu:~/[[ROOT]]/test$
-```
-
-Program #3
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1c(1,1)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-A
-student@ubuntu:~/[[ROOT]]/test$
-```
-
-Program #4
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1c(1,5)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-A
-B
-B
-B
-C
-student@ubuntu:~/[[ROOT]]/test$
-```
-
-----
-
-## raid1d
-
-### Instructions
-
-Write a function `Raid1d` that prints a **valid** rectangle of width `x` and of height `y`.
-
-The function must draw the rectangles as in the examples.
-
-`x` and `y` will always be positive numbers. Otherwise, the function should print nothing.
-
-### Expected function
-
-```go
-func Raid1d(x,y int) {
-
-}
-```
-
-### Usage
-
-Here are possible programs to test your function :
-
-Program #1
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1d(5,3)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-ABBBC
-B   B
-ABBBC
-student@ubuntu:~/[[ROOT]]/test$
-```
-
-Program #2
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1d(5,1)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-ABBBC
-student@ubuntu:~/[[ROOT]]/test$
-```
-
-Program #3
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1d(1,1)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-A
-student@ubuntu:~/[[ROOT]]/test$
-```
-
-Program #4
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1d(1,5)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-A
-B
-B
-B
-A
-student@ubuntu:~/[[ROOT]]/test$
-```
-
-----
-
-## raid1e
-
-### Instructions
-
-Write a function `Raid1e` that prints a **valid** rectangle of width `x` and of height `y`.
-
-The function must draw the rectangles as in the examples.
-
-`x` and `y` will always be positive numbers. Otherwise, the function should print nothing.
-
-### Expected function
-
-```go
-func Raid1e(x,y int) {
-
-}
-```
-
-### Usage
-
-Here are possible programs to test your function :
-
-Program #1
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1e(5,3)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-ABBBC
-B   B
-CBBBA
-student@ubuntu:~/[[ROOT]]/test$
-```
-
-Program #2
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1e(5,1)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-ABBBC
-student@ubuntu:~/[[ROOT]]/test$
-```
-
-Program #3
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1e(1,1)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-A
-student@ubuntu:~/[[ROOT]]/test$
-```
-
-Program #4
-
-```go
-package main
-
-import piscine ".."
-
-func main() {
-	piscine.Raid1e(1,5)
-}
-```
-
-And its output :
-
-```console
-student@ubuntu:~/[[ROOT]]/test$ go build
-student@ubuntu:~/[[ROOT]]/test$ ./test
-A
-B
-B
-B
-C
-student@ubuntu:~/[[ROOT]]/test$
-```
