--- subjects/raid1/audit/README.md
@@ -1,437 +1,0 @@
-#### raid1a
-
-##### Try running the function with the arguments: `"x=5 and y=3"`
-
-```
-o---o
-|   |
-o---o
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=5 and y=1"`
-
-```
-o---o
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=1 and y=1"`
-
-```
-o
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=1 and y=5"`
-
-```
-o
-|
-|
-|
-o
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=0 and y=0"`
-
-###### Does the function returns nothing?
-
-##### Try running the function with the arguments: `"x=-1 and y=6"`
-
-###### Does the function returns nothing?
-
-##### Try running the function with the arguments: `"x=6 and y=-1"`
-
-###### Does the function returns nothing?
-
-##### Try running the function with the arguments: `"x=20 and y=1"`
-
-```
-o------------------o
-```
-
-###### Does the function returns the value above?
-
-
-##### Try running the function with the arguments: `"x=10 and y=8"`
-
-```
-o--------o
-|        |
-|        |
-|        |
-|        |
-|        |
-|        |
-o--------o
-```
-
-###### Does the function returns the value above?
-
-#### raid1b
-
-##### Try running the function with the arguments: `"x=5 and y=3"`
-
-```
-/***\
-*   *
-\***/
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=5 and y=1"`
-
-```
-/***\
-```
-
-###### Does the function returns the value above?
-
-
-##### Try running the function with the arguments: `"x=1 and y=1"`
-
-```
-/
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=1 and y=5"`
-
-```
-/
-*
-*
-*
-\
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=0 and y=0"`
-
-###### Does the function returns nothing?
-
-##### Try running the function with the arguments: `"x=-1 and y=6"`
-
-###### Does the function returns nothing?
-
-##### Try running the function with the arguments: `"x=6 and y=-1"`
-
-###### Does the function returns nothing?
-
-##### Try running the function with the arguments: `"x=18 and y=6"`
-
-```
-/****************\
-*                *
-*                *
-*                *
-*                *
-\****************/
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=9 and y=3"`
-
-```
-/*******\
-*       *
-\*******/
-```
-
-###### Does the function returns the value above?
-
-#### raid1c
-
-##### Try running the function with the arguments: `"x=5 and y=3"`
-
-```
-ABBBA
-B   B
-CBBBC
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=5 and y=1"`
-
-```
-ABBBA
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=1 and y=1"`
-
-```
-A
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=1 and y=5"`
-
-```
-A
-B
-B
-B
-C
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=0 and y=0"`
-
-###### Does the function returns nothing?
-
-##### Try running the function with the arguments: `"x=-1 and y=6"`
-
-###### Does the function returns nothing?
-
-##### Try running the function with the arguments: `"x=6 and y=-1"`
-
-###### Does the function returns nothing?
-
-##### Try running the function with the arguments: `"x=13 and y=7"`
-
-```
-ABBBBBBBBBBBA
-B           B
-B           B
-B           B
-B           B
-B           B
-CBBBBBBBBBBBC
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=10 and y=15"`
-
-```
-ABBBBBBBBA
-B        B
-B        B
-B        B
-B        B
-B        B
-B        B
-B        B
-B        B
-B        B
-B        B
-B        B
-B        B
-B        B
-CBBBBBBBBC
-```
-
-###### Does the function returns the value above?
-
-#### raid1d
-
-##### Try running the function with the arguments: `"x=5 and y=3"`
-
-```
-ABBBC
-B   B
-ABBBC
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=5 and y=1"`
-
-```
-ABBBC
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=1 and y=1"`
-
-```
-A
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=1 and y=5"`
-
-```
-A
-B
-B
-B
-A
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=0 and y=0"`
-
-###### Does the function returns nothing?
-
-##### Try running the function with the arguments: `"x=-1 and y=6"`
-
-###### Does the function returns nothing?
-
-##### Try running the function with the arguments: `"x=6 and y=-1"`
-
-###### Does the function returns nothing?
-
-##### Try running the function with the arguments: `"x=3 and y=16"`
-
-```
-ABC
-B B
-B B
-B B
-B B
-B B
-B B
-B B
-B B
-B B
-B B
-B B
-B B
-B B
-B B
-ABC
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=7 and y=16"`
-
-```
-ABBBBBC
-B     B
-B     B
-B     B
-B     B
-B     B
-B     B
-B     B
-B     B
-B     B
-B     B
-B     B
-B     B
-B     B
-B     B
-ABBBBBC
-```
-
-###### Does the function returns the value above?
-
-#### raid1e
-
-##### Try running the function with the arguments: `"x=5 and y=3"`
-
-```
-ABBBC
-B   B
-CBBBA
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=5 and y=1"`
-
-```
-ABBBC
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=1 and y=1"`
-
-```
-A
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=1 and y=5"`
-
-```
-A
-B
-B
-B
-C
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=0 and y=0"`
-
-###### Does the function returns nothing?
-
-##### Try running the function with the arguments: `"x=-1 and y=6"`
-
-###### Does the function returns nothing?
-
-##### Try running the function with the arguments: `"x=6 and y=-1"`
-
-###### Does the function returns nothing?
-
-##### Try running the function with the arguments: `"x=21 and y=24"`
-
-```
-ABBBBBBBBBBBBBBBBBBBC
-B                   B
-B                   B
-B                   B
-B                   B
-B                   B
-B                   B
-B                   B
-B                   B
-B                   B
-B                   B
-B                   B
-B                   B
-B                   B
-B                   B
-B                   B
-B                   B
-B                   B
-B                   B
-B                   B
-B                   B
-B                   B
-B                   B
-CBBBBBBBBBBBBBBBBBBBA
-
-```
-
-###### Does the function returns the value above?
-
-##### Try running the function with the arguments: `"x=18 and y=8"`
-
-```
-ABBBBBBBBBBBBBBBBC
-B                B
-B                B
-B                B
-B                B
-B                B
-B                B
-CBBBBBBBBBBBBBBBBA
-
-```
-
-###### Does the function returns the value above?
